Add ffmpeg package and library names for ubuntu 15.04
Add the ffmpeg versions of the libraries because the libav versions available
on ubuntu 15.04 have not yet wrapped in VideoIO.

The ffmpeg libraries have been put first in the search order so that they'll be
found in preference even if the user also has the libav versions installed.
This avoids a VideoIO crash at runtime when loading libraries which haven't
been wrapped yet.