Increase efficiency of accessing members of struct pointers
The current code to access a member of a struct pointer uses a package function,
`fieldposition`, that has identical functionality to a function provided in
base, `fieldindex`. While this latter function is not exported, it is documented
and will hopefully remain a part of Julia's API. I imagine that the base
function is more performant than the implementation in `fieldposition`, which
compares each field name to the target field name in order to find the position
of a field. As such, I have replaced calls to `fieldposition` with calls to
`Base.fieldindex`, and deleted the function `fieldposition`.

Additionally, when setting a member of a struct pointer, the current
implementation calculates the pointer position, wraps it in a julia array with
`unsafe_wrap`, and then sets the first element of that array to the desired
value. The same result is achieved, perhaps more simply, by using
`unsafe_store!` and not using an intermediate array. I have therefore replaced
the call to `unsafe_wrap` and subsequent code with `unsafe_store!`.

While I do not know if the existing code causes a measurable decrease in
performance, using the base functions reduces the amount of code, and most
likely makes it faster.