Allow read! to work with regular arrays, not just PermutedDimsArrays
Commit eaf4dbc fixed a bug where FFMPEG's line size alignment requirements were
not met, but in doing so unintentionally restricted the types of Julia arrays
that would would work with VideoIO's `read!` from the original `VidArray =
Union{PermutedArray,Array}` to just `PermutedArrays`. While the default return
type of VideoIO's `read` is a `PermutedDimsArray`, this commit caused regular
arrays to throw an error when used with `read!`, and in effect changed
the (undocumented) API of `read!` between VideoIO versions v0.7.1 and v0.7.2.
Furthermore, the fixed `read!` that respects FFMPEG's SIMD alignemnt
requirements should work with either a normal `Array`, or instead a
`PermutedDimsArrays` view of that array; the memory layout of both types of
buffers are identical.

I have removed the line that caused the error when used with normal arrays: an
attempt to access the `parent` field of the buffer. I have also changed the
source of frame width and height information from the size of the permuted
buffer to the width and height information given by the transcoder. This switch
eliminates another bug when copying lines from the AVFrame managed by ffmpeg
into the buffer used by Julia client code when that second buffer is a regular
array. In addition to fixing this bug, the FFMPEG transcoder is a more
authoritative source of information about frame dimensions, and indeed no
checking was done to ensure that the height and width of the buffer matches the
height and width of the transcoder.

I have also eliminated the temporary variables `o_stride` and `o_size`, which
are never used.

I have added a bit of documentation to VideoIO's `read!` function, describing
the valid types of buffers for `read!`.

Finally, I have added tests to ensure `read!` works with both
`PermutedDimsArray`s as well as `Array`s. In the process, I have refactored some
code that would otherwise be repeated for these tests into helper functions.