Make some fields in `VideoReader` concrete types (#258)
Two fields of `VideoReader`, `framerate` and `aspect_ratio`, are both of type
`Rational`, an abstract type. In reality, the constructor of `VideoReader` sets
both of these fields to be type `Rational{Cint}`. By changing the type
definition to reflect that, it reduces type instability in code accessing
`VideoReader` fields, while not restricting the generality of `VideoReader`.