Add iterator interface to VideoReader objects
I have added an iterator interface to the `VideoReader` type, making collecting
all frames of a video as simple as `f = openvideo(fname); imgs = collect(f)`.
The iterator returns a new array for each frame, as I think mutating the same
array on subsequent calls to `iterate` would be surprising and lead to strange
behavior when calling functions such as `collect`. Note that the VideoReader
iterator is mutable, like Channel or Stateful iterators in Julia, and will
resume iteration where it was last stopped, instead of at the beginning. I have
added some text in the documentation explaining this behavior, and also added
some tests.