Fix bug when encoding Gray videos
I have been using this function with UInt8 or N0f8 el type, which worked fine.
If the el type is instead <:Gray, and `scanline_major = true`, then there would
be a method error. This was because `rawview` does not return a `StridedArray`.
I have replaced `rawview` with a simple function that suffices for this `Gray`
case, and produces a `StridedArray`.