Make NO_TRANSCODE VideoReaders work
Prior to this commit, there was a method error when using `NO_TRANSCODE`
`VideoReader`s because the type signature of the `NO_TRANSCODE` `retrieve!(...,
buf)` did not accept byte buffers, even though that is what the code requires.

Additionally, I have changed the API of the `NO_TRANSCODE` `retrieve` method to
return both the byte buffer, as well as the alignment of that buffer. The buffer
requires some non-zero alignment, and knowing this alignment is critical to
interpreting the contents of the byte buffer. While the alignment could be set
to 1 to be unsurprising to Julia users, users hoping to pass this buffer back to
FFMPEG would likely find anything other than 32 surprising. Instead of picking a
default for one group or the other, or implicitly making this alignment
parameter part of the API, I think returning the alignment with the buffer makes
the most sense, and additionally does not commit the API to a particular
alignment. This means `read(...)` will return the tuple `(buf, align)`, instead
of just the buffer, as is done in the `TRANSCODE` case. Additionally, I have
added optional `align` arguments to both `read` and `read!`, which defaults to
the VIDEOIO_ALIGN value that is used internally (32).