Return the alignment used for byte buffer for `retrieve_raw`
Since there is not an obvious alignment for byte buffers (e.g.
user-interpretable byte buffers might be 1-aligned, whereas byte buffers for
ffmpeg are often 32-aligned), I am returning the alignment used with
`retrieve_raw`, so that users get both the byte buffer and the alignment for
that buffer. This should allow the alignment used to not be part of the API of
VideoIO.jl.