Make NO_TRANSCODE VideoReader's work as expected
If a `VideoReader` is created with `transcode = false`, then calls to `read`
should return just a byte buffer with 1-aligned data. Additionally, the output
should be the pixel format of the video stream, not the destination pixel
format.

I have also added tests to make sure that `read_raw` and reading with `transcode
= false` works as expected.