Check that output buffer is both the right size and type
Currently, the size of the output buffer is checked, but not the type, before
reading the next frame from the video stream. If the output buffer is the wrong
type, then a method error will be thrown, but only after advancing the stream.

I have changed this behavior so that both the type and size are checked prior to
advancing the stream, so that a user will not 'skip' a frame if they supply the
wrong buffer to transfer data into.

I have also added internal `_retrieve!` methods that do not perform size and
type checking of the destination buffer, which is used if VideoIO has generated
the destination buffer (and is therefore assumed to be right).