Export utility functions needed to deal with raw buffers, add tests
I have added and exported a few utility functions that are necessary to
interpret the output of `read_raw`: `raw_pixel_format`, `out_bytes_size`, and `raw_frame_size`.

I have also changed size and eltype checks for `read!` and `read_raw!` to throw
`ArgumentError` instead of `ErrorException`.

I have expanded the tests to include testing that errors are thrown as expected,
and the video stream is not advanced, when incorrect destination buffers are
passed to `read!` and `read_raw!`. I have also added some test cases to make
sure that reading and seeking are consistent. I have additionally added tests of
the exported utility functions, such as `out_frame_size`. Finally, there were
some tests that did not work because they did not account for the frame buffer
being modified over successive calls to `read!`, resulting in a test that
trivially test if a buffer is equal to itself, rather than the intended test of
whether the fiftieth frame can reliably be read after seeking.

I have also added a test set that checks that monochrome videos can be read into
UInt8 buffers, and that the pixel values from a limited range YCbCr video are
expanded to the full 0-255 range. I have also tested that this behavior can be
disabled by changing the `target_colorspace_details` kwarg of `VideoReader`.