Support do-block syntax for openvideo, prepareencoder, open_video_out
I have added methods which accept a single function as the first argument to
`openvideo`, `prepareencoder`, and `open_video_out`. These functions will
automatically close the video stream at the end of the call, regardless if the
call was successful. This allows the use of 'do block' syntactic sugar similar
Base's `open`.

I have additionally made `encode_mux_video` use this do-block syntax to ensure
that the video stream is closed in the event of an error.