Make sure files and streams are closed during testing, and use tempdir
I have made it so that testing VideoIO does not require writing or modifying
files in the test directory. From what I can tell, Julia will not allow the
package contents (including test directory) to be modified at some point. I have
instead made it so that test generate files in the temporary directory, instead
of the test directory.

I have additionally used the function methods of `encodevideo` and `openvideo`
to make sure that their files are properly closed before use, and before the
file is deleted. This may help with testing problems on Windows, which
occasionally errors because files are in use when the test script attempts to
delete them.