Restore functionality of IO stream VideoReader
Somewhere in this PR the functionality of the IO stream VideoReader was deleted.
I have added the code back. However, there are no tests for this functionality,
so I am not sure that it currently works, nor am I sure that it works on master.

I will add a test of this functionality if I have time.