Close AVInput AVFormatContext immediately if `close`d
Instead of waiting for the GC to finalize a the `AVFormatContext` of a `AVInput`
to close the input file, I have changed the `close` function to immediately
close the `AVFormatContext`. This means the finalizer may attempt to close it
again, but I believe that `avformat_close_input` can handle null pointers.