Check that VideoWriter is open before attempting to encode
Currently `append_encode_mux!` does not check that a `VideoWriter` is open and
valid before attempting to access its memory. If it is not valid, then this will
result in a segfault. Since this is a user-facing function, I have added a check
to `append_encode_mux!` that will throw an `ErrorException` if `VideoWriter` is
closed or invalid. I have also added a test to ensure that an exception is
thrown, rather than triggering a segfault.